WarpedVRT does its thing in __init__ rather than start
